change apidoc generator location in route
