+++ apidocs/apidoc.html.json
@@ -1,1 +1,1 @@
-{"ApiSpecs":[{"HttpVerb":"GET","Path":"/v1/user/1","Calls":[{"Id":0,"CurrentPath":"/v1/user/1","MethodType":"GET","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r","Postman-Token":" c1848f43-22d9-ef8f-153b-563e99e7b8c1\r"},"CommonRequestHeaders":null,"ResponseHeader":{"Content-Type":"application/json; charset=UTF-8"},"RequestUrlParams":{},"RequestBody":"","ResponseBody":"{\"count\":1,\"error\":\"false\",\"result\":{\"id\":1,\"role\":\"user\",\"email\":\"lol@gits.id\",\"password\":\"$2a$10$OwPGDQh8LZpooxN39HXk2u8BD4rUs9CLvISYtIivB6SHNIMc2.7Au\",\"createdAt\":\"2018-10-04T07:23:42Z\",\"updatedAt\":\"2018-10-04T07:23:42Z\",\"profile\":{\"id\":0}},\"status\":200}","ResponseCode":200}]}]}
+{"ApiSpecs":[{"HttpVerb":"GET","Path":"/v1/user/1","Calls":[{"Id":0,"CurrentPath":"/v1/user/1","MethodType":"GET","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r","Postman-Token":" c1848f43-22d9-ef8f-153b-563e99e7b8c1\r"},"CommonRequestHeaders":null,"ResponseHeader":{"Content-Type":"application/json; charset=UTF-8"},"RequestUrlParams":{},"RequestBody":"","ResponseBody":"{\"count\":1,\"error\":\"false\",\"result\":{\"id\":1,\"role\":\"user\",\"email\":\"lol@gits.id\",\"password\":\"$2a$10$OwPGDQh8LZpooxN39HXk2u8BD4rUs9CLvISYtIivB6SHNIMc2.7Au\",\"createdAt\":\"2018-10-04T07:23:42Z\",\"updatedAt\":\"2018-10-04T07:23:42Z\",\"profile\":{\"id\":0}},\"status\":200}","ResponseCode":200}]},{"HttpVerb":"GET","Path":"/v1/user","Calls":[{"Id":0,"CurrentPath":"/v1/user","MethodType":"GET","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r","Postman-Token":" 06578646-97cf-084a-cc1d-e841cf992da8\r"},"CommonRequestHeaders":null,"ResponseHeader":{"Content-Type":"application/json; charset=UTF-8"},"RequestUrlParams":{},"RequestBody":"","ResponseBody":"{\"count\":3,\"error\":\"false\",\"result\":[{\"id\":1,\"role\":\"user\",\"email\":\"lol@gits.id\",\"createdAt\":\"2018-10-04T07:23:42Z\",\"updatedAt\":\"2018-10-04T07:23:42Z\",\"profile\":{\"id\":0}},{\"id\":2,\"role\":\"user\",\"email\":\"wahyu@gits.id\",\"createdAt\":\"2018-10-04T08:42:04Z\",\"updatedAt\":\"2018-10-04T08:42:04Z\",\"profile\":{\"id\":0}},{\"id\":3,\"role\":\"user\",\"email\":\"wahyu@gits.id\",\"createdAt\":\"2018-10-17T03:41:33Z\",\"updatedAt\":\"2018-10-17T03:41:33Z\",\"profile\":{\"id\":0}}],\"status\":200}","ResponseCode":200}]},{"HttpVerb":"POST","Path":"/v1/user/login","Calls":[{"Id":1,"CurrentPath":"/v1/user/login","MethodType":"POST","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r","Postman-Token":" 73f5ed09-8935-e572-27a5-871e101d9e43\r"},"CommonRequestHeaders":null,"ResponseHeader":{"Content-Type":"application/json; charset=UTF-8"},"RequestUrlParams":{},"RequestBody":"{\n\t\"email\" : \"wahyu@gits.id\",\n\t\"password\" : \"wahyu\"\n}","ResponseBody":"{\"error\":\"false\",\"message\":\"success login\",\"role\":\"user\",\"status\":200,\"token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6IndhaHl1QGdpdHMuaWQiLCJleHAiOjE1Mzk3Njc4MDAsImlkIjoyLCJyb2xlIjoidXNlciJ9.rX5hA2WLOgBY3Qm7oQkCFus3m-7R9LvTIzc8IW_aRYs\"}","ResponseCode":200}]}]}
